+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 **Feature - 10DLC api updates**
 - Update 10dlc api such as campaign, brand, profile and number link
 
+## [7.11.0](https://github.com/plivo/plivo-go/tree/v7.11.0) (2022-08-01)
+**Feature - Token creation**
+- `JWT Token Creation API` added functionality to create a new JWT token.
+
 ## [7.10.0](https://github.com/plivo/plivo-go/tree/v7.10.0) (2022-07-11)
 **Feature - STIR Attestation**
 - Add stir attestation param as part of Get CDR and Get live call APIs Response